HuntingSpiders {partykit}	R Documentation

Abundance of Hunting Spiders

Description

Abundances for 12 species of hunting spiders along with environmental predictors, all rated on a 0–9 scale.

Usage

data("HuntingSpiders")

Format

A data frame containing 28 observations on 18 variables (12 species abundances and 6 environmental predictors).

arct.lute

numeric. Abundance of species Arctosa lutetiana (on a scale 0–9).

pard.lugu

numeric. Abundance of species Pardosa lugubris (on a scale 0–9).

zora.spin

numeric. Abundance of species Zora spinimana (on a scale 0–9).

pard.nigr

numeric. Abundance of species Pardosa nigriceps (on a scale 0–9).

pard.pull

numeric. Abundance of species Pardosa pullata (on a scale 0–9).

aulo.albi

numeric. Abundance of species Aulonia albimana (on a scale 0–9).

troc.terr

numeric. Abundance of species Trochosa terricola (on a scale 0–9).

alop.cune

numeric. Abundance of species Alopecosa cuneata (on a scale 0–9).

pard.mont

numeric. Abundance of species Pardosa monticola (on a scale 0–9).

alop.acce

numeric. Abundance of species Alopecosa accentuata (on a scale 0–9).

alop.fabr

numeric. Abundance of species Alopecosa fabrilis (on a scale 0–9).

arct.peri

numeric. Abundance of species Arctosa perita (on a scale 0–9).

water

numeric. Environmental predictor on a scale 0–9.

sand

numeric. Environmental predictor on a scale 0–9.

moss

numeric. Environmental predictor on a scale 0–9.

reft

numeric. Environmental predictor on a scale 0–9.

twigs

numeric. Environmental predictor on a scale 0–9.

herbs

numeric. Environmental predictor on a scale 0–9.

Details

The data were originally analyzed by Van der Aart and Smeenk-Enserink (1975). De'ath (2002) transformed all variables to the 0–9 scale and employed multivariate regression trees.

Source

Package mvpart (currently archived, see https://CRAN.R-project.org/package=mvpart).

References

Van der Aart PJM, Smeenk-Enserink N (1975). Correlations between Distributions of Hunting Spiders (Lycosidae, Ctenidae) and Environmental Characteristics in a Dune Area. Netherlands Journal of Zoology, 25, 1–45.

De'ath G (2002). Multivariate Regression Trees: A New Technique for Modelling Species-Environment Relationships. Ecology, 83(4), 1103–1117.

Examples

## load data
data("HuntingSpiders", package = "partykit")

## fit multivariate tree for 12-dimensional species abundance
sptree <- ctree(arct.lute + pard.lugu + zora.spin + pard.nigr + pard.pull +
  aulo.albi + troc.terr + alop.cune + pard.mont + alop.acce + alop.fabr +
  arct.peri ~ herbs + reft + moss + sand + twigs + water, data = HuntingSpiders,
  teststat = "max", minsplit = 5)
plot(sptree, terminal_panel = node_barplot)

[Package partykit version 1.2-2 Index]