extree_fit {partykit}	R Documentation

Fit Extensible Trees.

Description

Basic infrastructure for fitting extensible trees.

Usage

extree_fit(data, trafo, converged, selectfun = ctrl$selectfun, splitfun = ctrl$splitfun, 
           svselectfun = ctrl$svselectfun, svsplitfun = ctrl$svsplitfun, partyvars, 
           subset, weights, ctrl, doFit = TRUE)

Arguments

data
an object of class extree_data, see extree_data.

trafo
a function with arguments subset, weights, info, estfun and object.

converged
a function with arguments subset, weights.

selectfun
an optional function for selecting variables.

splitfun
an optional function for selecting splits.

svselectfun
an optional function for selecting surrogate variables.

svsplitfun
an optional function for selecting surrogate splits.

partyvars
a numeric vector assigning a weight to each partitioning variable (z in extree_data.

subset
a sorted integer vector describing a subset.

weights
an optional vector of weights.

ctrl
control arguments.

doFit
a logical indicating if the tree shall be grown (TRUE) or not FALSE.

Details

This internal functionality will be the basis of implementations of other tree algorithms in future versions. Currently, only ctree relies on this function.

Value

An object of class partynode.

[Package partykit version 1.2-2 Index]