model_frame_rpart {partykit}	R Documentation

Model Frame Method for rpart

Description

A model.frame method for rpart objects.

Usage

model_frame_rpart(formula, ...)

Arguments

formula
an object of class rpart.

...
additional arguments.

Details

A model.frame method for rpart objects. Because it is no longer possible to overwrite existing methods, the function name is a little different here.

Value

A model frame.

[Package partykit version 1.2-12 Index]